+++ app/Console/Commands/ProBINDPushZones.php
@@ -17,6 +17,7 @@
 
 namespace App\Console\Commands;
 
+use App\Helpers\SFTP\SFTP;
 use App\Server;
 use App\Zone;
 use Carbon\Carbon;
@@ -24,8 +25,6 @@ use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\View;
 use phpseclib\Crypt\RSA;
-use phpseclib\Net\SFTP;
-use Setting;
 
 /**
  * Class ProBINDPushZones.
@@ -48,22 +47,9 @@ class ProBINDPushZones extends Command
      */
     protected $description = 'Generate and push zone files to DNS servers';
 
-    /**
-     * The local Storage path to be used.
-     *
-     * @var string
-     */
-    protected $localStoragePath;
-
-    /**
-     * Create a new command instance.
-     */
-    public function __construct()
-    {
-        $this->localStoragePath = storage_path('probind');
-
-        parent::__construct();
-    }
+    const BASEDIR = 'probind';
+    const CONFIG_BASEDIR = self::BASEDIR . DIRECTORY_SEPARATOR . 'configuration';
+    const ZONE_BASEDIR = self::BASEDIR . DIRECTORY_SEPARATOR . 'primary';
 
     /**
      * Execute the console command.
@@ -73,44 +59,53 @@ class ProBINDPushZones extends Command
     public function handle()
     {
         // Prepare local storage for file's creation
-        Storage::deleteDirectory($this->localStoragePath);
+        Storage::deleteDirectory(self::BASEDIR);
 
         // Generate a file containing zone's name that has been deleted
-        $deletedZones = Zone::onlyTrashed()->get();
-        $content = $this->generateDeletedZonesContent($deletedZones);
-        Storage::put($this->localStoragePath . '/configuration/deadlist', $content, 'private');
+        $deletedZones = Zone::onlyTrashed()
+            ->get();
+        $content = join("\n",
+            $deletedZones
+                ->pluck('domain')
+                ->all()
+        );
+        $path = self::CONFIG_BASEDIR . DIRECTORY_SEPARATOR . 'deadlist';
+        Storage::put($path, $content, 'private');
 
         // Generate one file for zone with its zone definition
-        $zonesToUpdate = Zone::withPendingChanges()->get();
+        $zonesToUpdate = Zone::withPendingChanges()
+            ->get();
         foreach ($zonesToUpdate as $zone) {
             $this->generateZoneFileForZone($zone);
         }
 
         // Now push files to servers using SFTP
-        $error = $this->handleAllServers();
+        if (false === $this->handleAllServers()) {
+            $this->error('Push updates completed with errors');
 
-        if (! $error) {
-            // Clear pending changes on zones and clear deleted ones
-            foreach ($zonesToUpdate as $zone) {
-                $zone->setPendingChanges(false);
-            }
-
-            foreach ($deletedZones as $zone) {
-                $zone->forceDelete();
-            }
+            return false;
         }
 
-        return $error;
+        // Clear pending changes on zones and clear deleted ones
+        foreach ($zonesToUpdate as $zone) {
+            $zone->setPendingChanges(false);
+        }
+
+        foreach ($deletedZones as $zone) {
+            $zone->forceDelete();
+        }
+
+        $this->info('Push updates completed successfully.');
+
+        return true;
     }
 
     /**
      * Returns the content of Deleted Zones File.
      *
-     * @param array $deletedZones
-     *
      * @return string
      */
-    public function generateDeletedZonesContent(array $deletedZones): string
+    public function generateDeletedZonesContent($deletedZones): string
     {
         $content = [];
         foreach ($deletedZones as $zone) {
@@ -123,16 +118,14 @@ class ProBINDPushZones extends Command
     /**
      * Creates a file with the zone definitions.
      *
-     * @param Zone $zone
+     * @param  Zone  $zone
      *
      * @return bool
      */
     public function generateZoneFileForZone(Zone $zone)
     {
-        $path = $this->localStoragePath . '/primary/' . $zone->domain;
-
         // Get default settings, we will use to render view
-        $defaults = Setting::all();
+        $defaults = setting()->all();
 
         // Get all Name Servers that had to be on NS records
         $nameServers = Server::where('ns_record', true)
@@ -147,17 +140,15 @@ class ProBINDPushZones extends Command
             ->orderBy('type')
             ->get();
 
-        // Put a header on generated file
-        $header = sprintf(";\n; This file has been automatically generated using ProBIND v3 on %s.\n;",
-            Carbon::now());
-        Storage::put($path, $header, 'private');
-
         // Create file content with a blade view
         $contents = view('templates.zone')
+            ->with('date', Carbon::now())
             ->with('defaults', $defaults)
             ->with('zone', $zone)
             ->with('servers', $nameServers)
             ->with('records', $records);
+
+        $path = self::ZONE_BASEDIR . DIRECTORY_SEPARATOR . $zone->domain;
 
         return Storage::append($path, $contents);
     }
@@ -167,7 +158,7 @@ class ProBINDPushZones extends Command
      *
      * @return bool
      */
-    public function handleAllServers()
+    public function handleAllServers(): bool
     {
         // Get servers with push updates capability
         $servers = Server::withPushCapability()->get();
@@ -177,18 +168,20 @@ class ProBINDPushZones extends Command
             return true;
         }
 
-        $pushedWithoutErrors = true;
+        $pushedWithErrors = false;
         foreach ($servers as $server) {
-            $pushedWithoutErrors &= (true === $this->handleServer($server));
+            if (false === $this->handleServer($server) && (false === $pushedWithErrors)) {
+                $pushedWithErrors = true;
+            }
         }
 
-        return $pushedWithoutErrors;
+        return ! $pushedWithErrors;
     }
 
     /**
      * Handle this command only for one Server.
      *
-     * @param Server $server
+     * @param  Server  $server
      *
      * @return bool
      */
@@ -199,22 +192,22 @@ class ProBINDPushZones extends Command
         // Create an array with files that need to be pushed to remote server
         $filesToPush = [
             [
-                'local' => $this->localStoragePath . '/configuration/' . $server->hostname . '.conf',
-                'remote' => Setting::get('ssh_default_remote_path') . '/configuration/named.conf',
+                'local' => self::CONFIG_BASEDIR . DIRECTORY_SEPARATOR . $server->hostname . '.conf',
+                'remote' => setting()->get('ssh_default_remote_path') . '/configuration/named.conf',
             ],
             [
-                'local' => $this->localStoragePath . '/configuration/deadlist',
-                'remote' => Setting::get('ssh_default_remote_path') . '/configuration/deadlist',
+                'local' => self::CONFIG_BASEDIR . DIRECTORY_SEPARATOR . 'deadlist',
+                'remote' => setting()->get('ssh_default_remote_path') . '/configuration/deadlist',
             ],
         ];
 
         if ($server->type == 'master') {
-            $localFiles = Storage::files($this->localStoragePath . '/primary/');
+            $localFiles = Storage::files(self::ZONE_BASEDIR);
             foreach ($localFiles as $file) {
                 $filename = basename($file);
                 $filesToPush[] = [
                     'local' => $file,
-                    'remote' => Setting::get('ssh_default_remote_path') . '/primary/' . $filename,
+                    'remote' => setting()->get('ssh_default_remote_path') . '/primary/' . $filename,
                 ];
             }
         }
@@ -225,13 +218,13 @@ class ProBINDPushZones extends Command
     /**
      * Create a file with DNS server configuration.
      *
-     * @param Server $server
+     * @param  Server  $server
      *
      * @return bool
      */
     public function generateConfigFileForServer(Server $server)
     {
-        $path = $this->localStoragePath . '/configuration/' . $server->hostname . '.conf';
+        $path = self::CONFIG_BASEDIR . DIRECTORY_SEPARATOR . $server->hostname . '.conf';
 
         // We allow specific templates for each server
         $templateFile = $this->getTemplateForConfigFile($server);
@@ -255,7 +248,7 @@ class ProBINDPushZones extends Command
     /**
      * Returns the template for rendering configuration file.
      *
-     * @param Server $server
+     * @param  Server  $server
      *
      * @return string
      */
@@ -272,53 +265,50 @@ class ProBINDPushZones extends Command
     /**
      * Push files to a Master server using SFTP.
      *
-     * @param Server $server
-     * @param array  $filesToPush
+     * @param  Server  $server
+     * @param  array  $filesToPush
      *
      * @return bool
      */
-    public function pushFilesToServer(Server $server, $filesToPush)
+    public function pushFilesToServer(Server $server, $filesToPush): bool
     {
-        // Get RSA private key in order to connect to servers
-        $privateSSHKey = new RSA();
-        $privateSSHKey->loadKey(Setting::get('ssh_default_key'));
-
         try {
-            $sftp = new SFTP($server->hostname, Setting::get('ssh_default_port'));
-        } catch (\Exception $e) {
-            $this->error('Can\'t connect to ' . $server->hostname . ': ' . $e->getMessage());
+            $this->info('Connecting to ' . setting()->get('ssh_default_user') . '@' . $server->hostname . ' (' . setting()->get('ssh_default_port') . ')...');
+
+            // Get RSA private key in order to connect to servers
+            $privateSSHKey = new RSA();
+            if (false === $privateSSHKey->loadKey(setting()->get('ssh_default_key'))) {
+                $this->error('Invalid RSA private key, configure it on the Settings page.');
+
+                return false;
+            }
+
+            $sftp = new SFTP($server->hostname, setting()->get('ssh_default_port'));
+            $sftp->authWithPublicKey(setting()->get('ssh_default_user'), $privateSSHKey);
+        } catch (\Throwable $e) {
+            $this->error('Connection to ' . $server->hostname . ' failed: ' . $e->getMessage());
 
             return false;
         }
 
-        if (! $sftp->login(Setting::get('ssh_default_user'), $privateSSHKey)) {
-            $this->error('Invalid SSH credentials for ' . $server->hostname);
-
-            return false;
-        }
-
-        $this->info('Connected successfully to ' . $server->hostname);
-
-        // Create Remote Folders, last argument is to be recursive
-        $sftp->mkdir(Setting::get('ssh_default_remote_path') . '/configuration', -1, true);
-        $sftp->mkdir(Setting::get('ssh_default_remote_path') . '/primary', -1, true);
+        $this->info('Connected successfully to ' . $server->hostname . '.');
 
         $totalFiles = count($filesToPush);
         $pushedFiles = 0;
         foreach ($filesToPush as $file) {
-            $contents = Storage::get($file['local']);
+            $this->info('Uploading file [' . $file['local'] . ' -> ' . $file['remote'] . '].');
 
-            if (! $sftp->put($file['remote'], $contents)) {
+            if (false === $sftp->put(Storage::path($file['local']), $file['remote'])) {
                 $this->error('File ' . $file['local'] . ' can\'t be uploaded to ' . $server->hostname);
                 continue;
             }
             $pushedFiles++;
         }
 
-        $this->info('Has been pushed ' . $pushedFiles . ' of ' . $totalFiles . ' files to ' . $server->hostname);
+        $this->info('It has been pushed ' . $pushedFiles . ' of ' . $totalFiles . ' files to ' . $server->hostname . '.');
         $sftp->disconnect();
 
         // Return true if all files has been pushed
-        return $totalFiles == $pushedFiles;
+        return $totalFiles === $pushedFiles;
     }
 }