+++ resources/views/templates/zone.blade.php
@@ -1,5 +1,8 @@
+;
+; This file has been automatically generated using ProBIND v3 on {{ $date }}.
+
 $ORIGIN {{ $zone->domain }}
-$TTL {{ $zone->getDefaultTTL() }}
+$TTL {{ $zone->getDefaultTTL }}
 
 {{ $zone->getSOARecord() }}